settings: Move update_user_custom_profile_fields function.
This is a prep commit to avoid import cycles when we
would use update_user_custom_profile_fields function
in custom_profile_fields_ui.ts in next commit.
